All Controller: added @ModelAttribute("")
That was the frinkinngy salution!! Got almighty, I thought I would not
succed! but Justice should prevail and Projects should be done! this is
why I'm blazing!

The annotation was necessary to be able to test the methods while still
having a working app. @RequestBody would not cut it somehow but this one
does!